:art: Drop 11 wp.full fill-value casts the dtype argument already pins
``wp.full(n, wp.int32(-1), dtype=wp.int32, ...)`` casts a fill value to the type the adjacent
``dtype=`` already names. CLAUDE.md section 1.3: a cast to the type a value already has is noise.

The tree spelled it both ways, and the majority was already correct -- nine of the ten
``wp.full(..., dtype=wp.float32)`` sites pass a bare ``math.inf`` or ``float("inf")``, so the cast
was the outlier rather than the convention. Verified identical before removing any: same values and
same ``.dtype`` on both devices, at rank 1 and rank 2, for the int and float forms.

Swept tree-wide rather than only across the two files under review -- the review named three sites
in ``remesh.py`` and ``holes.py``, and a grep for the shape finds eleven across five files, so
stopping at three would have left the split it was reported for. A re-scan after the first pass
found two more the initial pattern missed (a cast of a *variable* rather than a literal, and a
second ``wp.float32(math.inf)``), which is why the check below keys on the constructor generically:

    grep -rn 'wp\.full(.*wp\.\(int32\|float32\|int64\|uint32\|uint64\|float64\|bool\)(' triwarp/ \
      | grep 'dtype=wp\.'

That now returns nothing. No behaviour change; 623 tests across the five touched modules pass.

+++ triwarp/grouping.py
@@ -550,7 +550,7 @@ def first_occurrence_indices(
     n = int(inverse.shape[0])
     if n_unique is None:
         n_unique = int(tw.reduce.max(inverse)) + 1 if n > 0 else 0
-    first = wp.full(n_unique, wp.int32(n), dtype=wp.int32, device=device)
+    first = wp.full(n_unique, n, dtype=wp.int32, device=device)
     if n > 0 and n_unique > 0:
         wp.launch(
             kernel_grouping.scatter_first_occurrence, dim=n, inputs=[inverse, first], device=device
